+++ .gitignore
@@ -21,3 +21,4 @@
 go.work
 
 .env
+NOTE
